+++ apollo-portal/src/main/java/com/ctrip/framework/apollo/portal/spi/springsecurity/SpringSecurityUserService.java
@@ -16,55 +16,54 @@
  */
 package com.ctrip.framework.apollo.portal.spi.springsecurity;
 
-import com.google.common.collect.Lists;
-
 import com.ctrip.framework.apollo.core.utils.StringUtils;
 import com.ctrip.framework.apollo.portal.entity.bo.UserInfo;
 import com.ctrip.framework.apollo.portal.entity.po.UserPO;
 import com.ctrip.framework.apollo.portal.repository.UserRepository;
 import com.ctrip.framework.apollo.portal.spi.UserService;
 
-import java.util.Collections;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.User;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.provisioning.JdbcUserDetailsManager;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.CollectionUtils;
 
 import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
-
-import javax.annotation.PostConstruct;
 
 /**
  * @author lepdou 2017-03-10
  */
 public class SpringSecurityUserService implements UserService {
 
-  private PasswordEncoder encoder = new BCryptPasswordEncoder();
-  private List<GrantedAuthority> authorities;
+  private final List<GrantedAuthority> authorities = Collections
+      .unmodifiableList(Arrays.asList(new SimpleGrantedAuthority("ROLE_user")));
 
-  @Autowired
-  private JdbcUserDetailsManager userDetailsManager;
-  @Autowired
-  private UserRepository userRepository;
+  private final PasswordEncoder passwordEncoder;
 
-  @PostConstruct
-  public void init() {
-    authorities = new ArrayList<>();
-    authorities.add(new SimpleGrantedAuthority("ROLE_user"));
+  private final JdbcUserDetailsManager userDetailsManager;
+
+  private final UserRepository userRepository;
+
+  public SpringSecurityUserService(
+      PasswordEncoder passwordEncoder,
+      JdbcUserDetailsManager userDetailsManager,
+      UserRepository userRepository) {
+    this.passwordEncoder = passwordEncoder;
+    this.userDetailsManager = userDetailsManager;
+    this.userRepository = userRepository;
   }
 
   @Transactional
   public void createOrUpdate(UserPO user) {
     String username = user.getUsername();
 
-    User userDetails = new User(username, encoder.encode(user.getPassword()), authorities);
+    User userDetails = new User(username, passwordEncoder.encode(user.getPassword()), authorities);
 
     if (userDetailsManager.userExists(username)) {
       userDetailsManager.updateUser(userDetails);
@@ -121,12 +120,6 @@ public class SpringSecurityUserService implements UserService {
       return Collections.emptyList();
     }
 
-    List<UserInfo> result = Lists.newArrayList();
-
-    result.addAll(users.stream().map(UserPO::toUserInfo).collect(Collectors.toList()));
-
-    return result;
+    return users.stream().map(UserPO::toUserInfo).collect(Collectors.toList());
   }
-
-
 }