+++ scripts/apollo-on-kubernetes/db/portal-db/apolloportaldb.sql
@@ -298,7 +298,7 @@ DROP TABLE IF EXISTS `Users`;
 CREATE TABLE `Users` (
   `Id` int(10) unsigned NOT NULL AUTO_INCREMENT COMMENT '自增Id',
   `Username` varchar(64) NOT NULL DEFAULT 'default' COMMENT '用户登录账户',
-  `Password` varchar(64) NOT NULL DEFAULT 'default' COMMENT '密码',
+  `Password` varchar(512) NOT NULL DEFAULT 'default' COMMENT '密码',
   `UserDisplayName` varchar(512) NOT NULL DEFAULT 'default' COMMENT '用户名称',
   `Email` varchar(64) NOT NULL DEFAULT 'default' COMMENT '邮箱地址',
   `Enabled` tinyint(4) DEFAULT NULL COMMENT '是否有效',